Reorganizar la estructura del panel de administración y agregar configuraciones para modelos de usuario, perfil de prestador y solicitudes de prestador
