+++ servic/admin/user_admin.py
@@ -1,0 +1,46 @@
+from django.contrib import admin
+from django.contrib.auth.admin import UserAdmin
+from ..models import User
+
+
+# Configuración personalizada para el modelo User
+class CustomUserAdmin(UserAdmin):
+    list_display = (
+        "email",
+        "username",
+        "first_name",
+        "last_name",
+        "user_type",
+        "is_staff",
+        "is_active",
+    )
+    list_filter = ("user_type", "is_staff", "is_active")
+    search_fields = ("email", "username", "first_name", "last_name")
+    ordering = ("email",)
+    fieldsets = (
+        (None, {"fields": ("email", "password")}),
+        ("Información Personal", {"fields": ("first_name", "last_name")}),
+        (
+            "Permisos",
+            {
+                "fields": (
+                    "user_type",
+                    "is_active",
+                    "is_staff",
+                    "is_superuser",
+                    "groups",
+                    "user_permissions",
+                )
+            },
+        ),
+        ("Fechas Importantes", {"fields": ("last_login", "date_joined")}),
+    )
+    add_fieldsets = (
+        (
+            None,
+            {
+                "classes": ("wide",),
+                "fields": ("email", "password1", "password2", "user_type"),
+            },
+        ),
+    )
